refactor: move model interface data to models.types.ts
