+++ packages/x-adapter-platform/src/schemas/next-query.schema.ts
@@ -1,6 +1,6 @@
 import { createMutableSchema, Schema } from '@empathyco/x-adapter-next';
 import { NextQuery } from '@empathyco/x-types';
-import { PlatformNextQuery } from '../types/response.types';
+import { PlatformNextQuery } from '../types/models.types';
 
 export const nextQuerySchema: Schema<PlatformNextQuery, NextQuery> = {
   query: 'query',
