+++ packages/x-adapter-platform/src/schemas/redirection.schema.ts
@@ -1,7 +1,7 @@
 import { createMutableSchema, Schema } from '@empathyco/x-adapter-next';
 import { Redirection, TaggingInfo } from '@empathyco/x-types';
 import { getTaggingInfoFromUrl } from '@empathyco/x-utils';
-import { PlatformRedirection } from '../types/response.types';
+import { PlatformRedirection } from '../types/models.types';
 
 const redirectionSchema: Schema<PlatformRedirection, Redirection> = {
   id: 'id',
