+++ packages/x-adapter-platform/src/types/index.ts
@@ -1,3 +1,4 @@
 export * from './platform-adapter.types';
 export * from './response.types';
 export * from './request.types';
+export * from './models.types';
