+++ packages/x-adapter-platform/src/types/response.types.ts
@@ -9,23 +9,23 @@ import {
   TaggingInfo
 } from '@empathyco/x-types';
 
+import {
+  Contentable,
+  PlatformBanner,
+  PlatformCatalog,
+  PlatformNextQuery,
+  PlatformPromoted,
+  PlatformRedirection,
+  PlatformRelatedTag,
+  PlatformResult,
+  PlatformSuggestion
+} from './models.types';
+
 export interface PlatformSearchResponse {
   banner: Contentable<PlatformBanner>;
   catalog: PlatformCatalog;
   direct: Contentable<PlatformRedirection>;
   promoted: Contentable<PlatformPromoted>;
-}
-export interface Contentable<T> {
-  content: T[];
-}
-
-export interface PlatformCatalog extends Contentable<PlatformResult> {
-  facets: PlatformFacet[];
-  numFound: number;
-  spellchecked: string;
-  tagging: {
-    query: string;
-  };
 }
 
 export interface SearchResponse {
@@ -39,75 +39,11 @@ export interface SearchResponse {
   queryTagging: TaggingInfo;
 }
 
-export interface PlatformResult {
-  name: string;
-  averageRating?: number;
-  id: string;
-  images?: string[];
-  image: string;
-  price: number;
-  sku?: string;
-  url: string;
-  tagging: PlatformTagging;
-}
-
-export interface PlatformFacet {
-  facet: string;
-  values: PlatformFacetFilter[];
-}
-
-export interface PlatformFacetFilter {
-  count: number;
-  filter: string;
-  id: string;
-  value: string;
-}
-
-export interface PlatformHierarchicalFilter extends PlatformFacetFilter {
-  children: PlatformFacet;
-}
-
-export interface PlatformTagging {
-  add2cart: string;
-  checkout: string;
-  click: string;
-}
-
-export interface PlatformPromoted {
-  id: string;
-  title: string;
-  url: string;
-  image_url: string;
-  tagging?: {
-    query: string;
-  };
-}
-export interface PlatformBanner {
-  id: string;
-  title: string;
-  url: string;
-  image_url: string;
-  tagging?: {
-    query: string;
-  };
-}
-export interface PlatformRedirection {
-  id: string;
-  url: string;
-  tagging?: {
-    click: string;
-  };
-}
-
 export interface PlatformEmpathizeResponse {
   topTrends: {
     content: PlatformSuggestion[];
     spellcheck?: string;
   };
-}
-
-export interface PlatformSuggestion {
-  title_raw: string;
 }
 
 export interface EmpathizeResponse {
@@ -136,12 +72,6 @@ export interface PlatformNextQueriesResponse {
   };
 }
 
-export interface PlatformNextQuery {
-  query: string;
-  source: 'ORGANIC' | 'CURATED';
-  position: number;
-}
-
 export interface RelatedTagsResponse {
   relatedTags: RelatedTag[];
 }
@@ -151,13 +81,6 @@ export interface PlatformRelatedTagsResponse {
     relatedtags: PlatformRelatedTag[];
   };
   status: number;
-}
-
-export interface PlatformRelatedTag {
-  query: string;
-  tag: string;
-  source: 'ORGANIC' | 'CURATED';
-  position: number;
 }
 
 export interface PlatformSkuSearchResponse {
